lab/ft_transcendence/backend/requirements.txt
asgiref==3.8.1
certifi==2020.6.20
chardet==4.0.0
Django==4.2.16
django-cors-headers==4.4.0
django-formtools==2.5.1
django-otp==1.5.4
django-phonenumber-field==8.0.0
django-two-factor-auth==1.17.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
httpie==2.2.0
idna==2.10
importchecker==3.0
phonenumbers==8.13.45
Pygments==2.7.1
PyJWT==2.9.0
pypng==0.20220715.0
qrcode==7.4.2
requests==2.25.1
six==1.16.0
sqlparse==0.5.1
typing-extensions==4.12.2
urllib3==1.26.5
